Cart (#42)
* added base component for cart and updatd order model to allow null fields for open orders

* updated cart route to send back only the order id and the items in the cart

* added images to shopping cart view

* added quantity field to shopping cart view

* connected rendered cart to cart stored on state

* cart rendering for logged in user from database

* ability to get cart from local storage for logged out users

* reformatted total to reflect current total

* fixed bug to render total

* added class method for order to get cart
